Backend/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
browser-cookie3==0.19.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.2.1
executing==0.8.3
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.3
intel-openmp==2021.4.0
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.2.2
jupyter-core==4.10.0
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
libretranslatepy==2.1.1
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
mtranslate==1.8
multidict==6.0.5
mysql-connector-python==9.1.0
namex==0.0.8
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==21.3
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
prompt-toolkit==3.0.29
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.0.8
pythainlp==5.0.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-gemini-api==2.4.12
pytz==2024.1
pywin32==303
PyYAML==6.0.1
pyzmq==22.3.0
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==56.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.2.0
sympy==1.13.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.1
tqdm==4.66.4
traitlets==5.1.1
transformers==4.42.4
translate==3.6.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.5
Werkzeug==3.0.3
wheel==0.43.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
